Refactor to split npm and dart
