+++ .github/workflows/ci.yml
@@ -28,7 +28,7 @@ jobs:
         fi
 
     - name: Run go vet
-      run: go vet ./...
+      run: go vet -all ./...
 
     - name: Run tests
       run: go test -v ./...
